feat(components): splitting code into components
