+++ src/shows/episodes/EpisodeDetailsPage.tsx
@@ -1,19 +1,10 @@
-import { Button } from "@mui/material";
 import React from "react";
-import { Link } from "react-router-dom";
 import EpisodeDetailsCard from "./EpisodeDetailsCardComponent";
 
 // interface IEpisodeDetailsPageProps {}
 
 const EpisodeDetailsPage: React.FC = () => {
-  return (
-    <>
-      <EpisodeDetailsCard />
-      <Link to="/">
-        <Button>Back to List</Button>
-      </Link>
-    </>
-  );
+  return <EpisodeDetailsCard />;
 };
 
 export default EpisodeDetailsPage;
